# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Provision Civils - Construction Job Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recycle-bin:
    get:
      operationId: listRecycleBin
      tags: [recycle]
      summary: List all soft-deleted items (admin only)
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [job, client, employee, invoice, photo]
      responses:
        "200":
          description: Deleted items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecycleBinItem"
  /recycle-bin/{type}/{id}/restore:
    post:
      operationId: restoreRecycleBinItem
      tags: [recycle]
      summary: Restore a soft-deleted item (admin only)
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum: [job, client, employee, invoice, photo]
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Item restored
  /recycle-bin/{type}/{id}:
    delete:
      operationId: permanentlyDeleteItem
      tags: [recycle]
      summary: Permanently delete a soft-deleted item (admin only)
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
            enum: [job, client, employee, invoice, photo]
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Permanently deleted
components:
  schemas:
...
    RecycleBinItem:
      type: object
      required: [id, type, name, deletedAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [job, client, employee, invoice, photo]
        name:
          type: string
        deletedAt:
          type: string
        metadata:
          nullable: true